refactor:시큐리티 허용 API추가,주석 변경,장르 패키지 이동
