+++ src/main/java/com/example/webduck/webtoon/repository/WebtoonRepository.java
@@ -16,7 +16,7 @@ public interface WebtoonRepository extends JpaRepository<Webtoon, Long> {
 
     List<Webtoon> findWebtoonByPlatform(Platform platform);
 
-    // 장르에 해당하는 웹툰 찾기
+    // 단건 장르에 해당하는 웹툰 찾기
     // TODO : 페이징 적용된다면 QueryDSL 로 변경해야 함
     @Query("SELECT w "
         + "FROM Webtoon w "
@@ -24,5 +24,4 @@ public interface WebtoonRepository extends JpaRepository<Webtoon, Long> {
         + "INNER JOIN Genre g ON wg.genre.id = g.id "
         + "WHERE g.name = :name")
     List<Webtoon> findByWebtoonsGenreName(@Param("name") String name);
-
 }
