+++ src/main/resources/static/js/webtoon/index/genreWebtoonList.js
