+++ src/main/resources/static/js/webtoon/index/platformWebtoonList.js
